Load up new config file to which we're uploading
Adds three command options:
     --update-conf [new config file]
     --move-content : simpy move the files instead of copying message
     --restore-update [aborted update dir]

Logic:
Loads the new config file the same way it's done for regular config.
If --restore-update is provided, check that the dir exists.  Set the
same command options for the new config file.  Initialize some vars
that will be used in the update process.

Discussion:
I'm thinking of completly changing the way the update process takes
its input. Instead of loading a new config file, this could be done in
a single config files with parameters for updating to new nametrans,
folderfilter and encoding. This could could open new possibilities to
improve performence of the updating process. It could also make the
--restore-update cmd option easier to use, since we can record its dir
in the config file.

Signed-off-by: Hubert Pineault <hpineault@riseup.net>