Invoke updating process
The update process is invoked through init.__sync() method.

Logic:
The idea is not to duplicate sig_handler and UI initializations.
Although, I think it should be rewritten in a distinct methode that
would called by run(). The sync process and the update process
shouldn't mixed.

Discussion:
I'm thinking of two ways to improve. Either write a new method that
would deal with sig_handler and be called by both sync and update
processes. Or, simply copy __sync() code and adapt it for the update
process.

New method:
    _newactiveaccounts()

The method is almost a copy/paste of _get_activeaccounts(). We first
make the list of accounts in the new config file. Then we check that
they exist in the old config file. It returns only accounts found in
both config files.  I think this code should be included in another
method so that we use _get_activeaccounts() and then test config's
files accounts correspondance. (Note that the whole war the update
process takes its input through files should probably be re-thought)

Signed-off-by: Hubert Pineault <hpineault@riseup.net>